+++ tests/Initialize.client.lua
@@ -2,4 +2,4 @@ local ReplicatedStorage = game:GetService("ReplicatedStorage")
 local fish = require(ReplicatedStorage.Packages.fish).Client
 
 fish.controllerDeep(script.Parent.Client.Controllers)
-fish.start()
+fish.start(true)